Move `@tractor_test` into `conftest.py`
